% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invarianceCutoffs.R
\name{invarianceCutoffs}
\alias{invarianceCutoffs}
\title{Fit Measure Cutoffs for Invariance Testing in Multigroup SEM}
\usage{
invarianceCutoffs(input = NULL, ...)
}
\arguments{
\item{input}{An object of the class \code{ezCutoffs}. Result of an \code{ezCutoffs} analysis of the configural model.}

\item{...}{Additional arguments to pass to \link[lavaan]{lavaan}. Ellipsis arguments already present in the call to \code{ezCutoffs} ought not to be repeated, since this can cause conflicts.}
}
\value{
An object of the class invarianceCutoffs, containing function calls, simulation parameters, fit distributions, the empirical model, and a summary.
For ease of analysis, it is further inspectable via \code{print}, \code{summary}, \code{plot}, and \code{\link{compareFit}}
}
\description{
A convenience function for determining cutoff values in multigroup invariance testing. Increasingly constrained models (cf. Meredith, 1993 <doi:10.1007/BF02294825>) are specified and compared in randomly generated data sets. Only the result of \code{\link{ezCutoffs}} is required as input.
}
\details{
\code{group} needs to be specified during \code{ezCutoffs()} analysis.\cr\cr
\code{group.equal} should not be specified because the function automatically tests configural, metric, scalar, and strict models, \code{group.partial} is valid. \cr\cr
Most other arguments should be specified in \code{ezCutoffs()} to yield accurate comparisons.
}
\examples{
## model specification examples

# simple uni-factorial model
model1 <- "F1 =~ a1 + a2 + a3 + a4 + a5"

out <- ezCutoffs(model1, n_obs = c(200, 200, 200), n_rep = 10, group = "group", n_cores = 1)

## function call
out <- invarianceCutoffs(input = out)

## retrieve output
print(out)
summary(out)

}
\references{
Meredith, W. (1993). Measurement invariance, factor analysis and factorial invariance. Psychometrika, 58(4), 525-543. doi: https://doi.org/10.1007/BF02294825
}
\seealso{
\code{\link{ezCutoffs}}
}
